+++ converter/oldChroma.ts
@@ -48,7 +48,7 @@ export function ogChromaToChromaV2<T extends IWrapDifficulty>(
    for (const ev of events) {
       let noChromaColor = false;
       if (ev.value >= 2000000000) {
-         currentColor[ev.type] = oldChromaColorConvert(ev.value) as ColorArray;
+         currentColor[ev.type] = oldChromaColorConvert(ev.value);
       }
       if (!currentColor[ev.type]) {
          noChromaColor = true;