+++ extensions/chroma/setColor.ts
@@ -4,7 +4,7 @@ import {
    SetColorOptions,
    SetColorRangeOptions,
 } from './types/colors.ts';
-import { convertColorType, interpolateColor } from '../../utils/colors.ts';
+import { convertColorType, lerpColor } from '../../utils/colors.ts';
 import { normalize } from '../../utils/math.ts';
 import { IChromaEventLight } from '../../types/beatmap/v3/custom/chroma.ts';
 import { settings } from './settings.ts';
@@ -45,13 +45,7 @@ export function setColorGradient(objects: IChromaObject[], options: SetColorGrad
    const endTime = objects.at(-1)!.time + opt.offsetEnd;
    objects.forEach((obj) => {
       const norm = normalize(obj.time, startTime, endTime);
-      const color = interpolateColor(
-         opt.colorStart,
-         opt.colorEnd,
-         norm,
-         opt.colorType,
-         opt.easingColor,
-      );
+      const color = lerpColor(opt.colorStart, opt.colorEnd, norm, opt.colorType, opt.easingColor);
       (obj.customData as IChromaEventLight).color = color;
    });
 }
@@ -70,7 +64,7 @@ export function setColorRandom(objects: IChromaObject[], options: SetColorRangeO
       if (objects[i].time > prevTime + 0.001) {
          random = Math.random();
       }
-      const color = interpolateColor(opt.color1, opt.color2, random, opt.colorType);
+      const color = lerpColor(opt.color1, opt.color2, random, opt.colorType);
       objects[i].customData._color = color;
       prevTime = objects[i].time;
    }